requirements.txt
beautifulsoup4==4.11.1
Cython==0.29.31
datasets==2.4.0
dateutils==0.6.12
einops==0.5.0
evaluate==0.2.2
h5py==3.7.0
hpargparse==0.14.0
hpman==0.13.0
huggingface-hub==0.9.1
lmdb==1.3.0
loguru==0.7.0
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mmcv==0.2.14
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
netron==6.7.8
networkx==2.6.3
nltk==3.7
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.21.6
omegaconf==2.2.3
onnx==1.13.1
onnx-simplifier==0.4.19
onnxruntime==1.14.1
opencv-python==4.6.0.66
pandas==1.3.5
pandocfilters==1.5.0
pathlib==1.0.1
pathtools==0.1.2
Pillow==9.2.0
protobuf==3.20.3
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pytz==2022.2.1
randaugment==1.0.2
reFILE==0.4.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
sacrebleu==2.2.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.9.0
setproctitle==1.3.0
shortuuid==1.0.9
sklearn==0.0.post1
smart-open==6.0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
starsessions==1.3.0
subword-nmt==0.3.8
sympy==1.10.1
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
terminado==0.17.1
terminaltables==3.1.10
Theano==1.0.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.11
tinycss2==1.2.1
tokenizers==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.11.4
torchsummary==1.5.1
torchtext==0.12.0
torchvision==0.12.0+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.65.0
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.22.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
ujson==5.7.0
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.21.1
uvloop==0.17.0
vidaug @ git+https://github.com/okankop/vidaug@1c1ddf2640fe4a9171267d64ae5e3bd70c24d54a
visualizer==0.0.1
wandb==0.12.21
watchfiles==0.19.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.1
widgetsnbextension==3.6.4
wrapt==1.14.1
xdg==5.1.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
